SwiGLU算子

+++ examples/activation/swi_glu.py
@@ -38,7 +38,7 @@ def swi_glu(M, N, block_M, block_N, dtype="float"):
                 T.barrier_all()
                 T.copy(A[bx * block_M + vid * block_M // VEC_NUM, by * block_N + N // 2], a1_ub)
                 T.barrier_all()
-                T.tile.swi_glu(b_ub, a0_ub, a1_ub, 1.0, block_M // VEC_NUM * block_N)
+                T.tile.swi_glu(b_ub, a0_ub, a1_ub, T.float32(1.0), block_M // VEC_NUM * block_N)
                 T.barrier_all()
                 T.copy(b_ub, B[bx * block_M + vid * block_M // VEC_NUM, by * block_N])
